Fixed issues with Strong or Wise cards while not in a drive form crashing the game. Turns out I wasnt checking if the stance to check was actually a driveform and not any other stance.
Fixed Combo cards triggering their own combo effect when no other combo cards were played. The code I had used to check if the previous card was a combo card only worked because it was in the card itself. It would effectively check for combo before the card itself was played. Now that all of those interactions have been moved into game actions, the original code would look at the current card in play. That has been changed.

The drive tags were also not working, so none of the strong or wise cards would ever actually trigger their given effect. It looks like I was using the wrong interface/built-in method for the object that holds the tags. I implemented the correct solution and those cards will trigger the respective effects when appropriate.

Also finally updated the README.md to be relevant to the mod rather than the generic README.md that was copied from the template I used.